feat: add fetch from reddit
To get things started, let's add a polyfill for fetch and just run
the fetch command to see what we get back, store it in the component
state and display the number of posts.

To continue development after this, it's time to add some tests that
mock the data so we can begin asserting what our component should
show.