feat: update design and components
The Reddit component is now split into Posts and Post components
which show the actual post data retrieved from the subreddit.

The subreddit is stored in the state for an upcoming feature of
choosing a subreddit to load from.

Images are shown alongside the title, with some initial styling
applied.